refactor: move AddressTranslator to new module
- Extract AddressTranslator trait and related code to a new addr_translator module
- Remove accessor module and its contents
- Update imports in lib.rs to use the new addr_translator module
